requirements_stweb.txt
# ST-WebAgentBench Complete Dependencies
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pdfplumber>=0.7.0
PyYAML>=6.0
anthropic>=0.7.0
openai-agents>=0.2.4
fastmcp==2.5.2

# ST-WebAgentBench specific dependencies
browsergym==v0.7.0
PyPDF2>=3.0.0
gymnasium>=0.29.1
tenacity>=8.0.0
nltk>=3.9

# Testing dependencies
pytest>=7.1.2
pytest-asyncio>=0.21.1

# Additional utilities
tqdm>=4.66.3
loguru>=0.7.0
rich>=13.0.0 